+++ packages/perseus/src/widget-ai-utils/matrix/matrix-ai-utils.ts
@@ -13,16 +13,16 @@ export type MatrixPromptJSON = {
 };
 
 export const getPromptJSON = (
-    renderProps: React.ComponentProps<typeof matrix.widget>,
+    widgetData: React.ComponentProps<typeof matrix.widget>,
 ): MatrixPromptJSON => {
     return {
         type: "matrix",
         options: {
-            height: renderProps.matrixBoardSize[0],
-            width: renderProps.matrixBoardSize[1],
+            height: widgetData.matrixBoardSize[0],
+            width: widgetData.matrixBoardSize[1],
         },
         userInput: {
-            answerRows: renderProps.userInput.answers,
+            answerRows: widgetData.userInput.answers,
         },
     };
 };