+++ packages/perseus/src/widgets/group/group.tsx
@@ -12,7 +12,6 @@ import classNames from "classnames";
 import * as React from "react";
 
 import {PerseusI18nContext} from "../../components/i18n-context";
-import * as Changeable from "../../mixins/changeable";
 import {ApiOptions} from "../../perseus-api";
 import Renderer from "../../renderer";
 import {
@@ -23,7 +22,6 @@ import {getPromptJSON as _getPromptJSON} from "../../widget-ai-utils/group/group
 
 import type {
     APIOptions,
-    ChangeFn,
     FocusPath,
     Widget,
     WidgetExports,
@@ -63,11 +61,6 @@ class Group extends React.Component<Props> implements Widget {
         // the group with the correct number.
         this.forceUpdate();
     }
-
-    change: ChangeFn = (...args) => {
-        // eslint-disable-next-line import/no-deprecated
-        return Changeable.change.apply(this, args);
-    };
 
     getPromptJSON(): GroupPromptJSON {
         return _getPromptJSON(this.rendererRef?.getPromptJSON());
@@ -130,17 +123,6 @@ class Group extends React.Component<Props> implements Widget {
             this.props.widgetId,
         );
 
-        // This is a little strange because the id of the widget that actually
-        // changed is going to be lost in favor of the group widget's id. The
-        // widgets prop also wasn't actually changed, and this only serves to
-        // alert our renderer (our parent) of the fact that some interaction
-        // has occurred.
-        const onInteractWithWidget = (id) => {
-            if (this.rendererRef) {
-                this.change("widgets", this.rendererRef.props.widgets);
-            }
-        };
-
         // TODO(mdr): Widgets inside this Renderer are not discoverable through
         //     the parent Renderer's `findWidgets` function.
         return (
@@ -166,7 +148,6 @@ class Group extends React.Component<Props> implements Widget {
                     findExternalWidgets={this.props.findWidgets}
                     reviewMode={this.props.reviewMode}
                     showSolutions={this.props.showSolutions}
-                    onInteractWithWidget={onInteractWithWidget}
                     linterContext={this.props.linterContext}
                     strings={this.context.strings}
                 />